refactor(analytics): split economics scatter modules
