+++ frontend/src/types.ts
@@ -763,57 +763,6 @@ export interface RateSpreadBucket {
   borrower_count: number;
 }
 
-export interface EquitySpreadPoint {
-  borrower_id: string;
-  display_name: string;
-  segment: string;
-  state: string;
-  equity_pct: number;
-  rate_spread_bps: number;
-  opportunity_score: number;
-  score_band?: 'high' | 'med' | 'low' | null;
-  in_the_money?: boolean | null;
-}
-
-export interface EquitySpreadViewport {
-  equity_min: number;
-  equity_max: number;
-  spread_min: number;
-  spread_max: number;
-}
-
-export interface EquitySpreadBin {
-  equity_bin_pct: number;
-  spread_bin_bps: number;
-  borrower_count: number;
-  mean_opportunity_score: number;
-  in_the_money_borrowers: number;
-}
-
-export interface EquitySpreadOverview {
-  bins: EquitySpreadBin[];
-  total_borrowers: number;
-  equity_bin_pct: number;
-  spread_bin_bps: number;
-  equity_domain_min: number;
-  equity_domain_max: number;
-  spread_domain_min: number;
-  spread_domain_max: number;
-  source_table: string;
-  refreshed_at?: string | null;
-}
-
-export interface EquitySpreadPointsResponse {
-  points: EquitySpreadPoint[];
-  total_matching: number;
-  showing: number;
-  point_cap: number;
-  truncated: boolean;
-  viewport: EquitySpreadViewport;
-  source_table: string;
-  refreshed_at?: string | null;
-}
-
 export interface TopBorrowerAnalyticsRow {
   borrower_id: string;
   display_name: string;
@@ -828,7 +777,7 @@ export interface TopBorrowerAnalyticsRow {
 
 export interface EconomicsAnalyticsResponse {
   rate_spread_histogram: RateSpreadBucket[];
-  equity_spread: EquitySpreadOverview;
+  equity_spread: import('./types/economicsScatter').EquitySpreadOverview;
   top_borrowers: TopBorrowerAnalyticsRow[];
 }
 
@@ -930,3 +879,5 @@ export type {
   LoanOfficerAssignment,
   SalesTeamMember,
 } from './types/loanOfficer';
+
+export type * from './types/economicsScatter';
